Resolving some issues with jest globals and eslint
My linter was choking on the jest globals feature, and none of the recommended
fixes online worked. Instead of fighting with it anymore, I opted to remove it
and use explicit imports in our tests. With this setup my environment works
flawlessly again. Let me know if this change doesn't work for you.